fixed selected intern bug, now interns are only excluded if they are deleted completely
